+++ Runtime/Scripts/PrimitiveCreateContext.cs
@@ -31,15 +31,22 @@ namespace GLTFast
 
         public JobHandle jobHandle;
         int[][] m_Indices;
+        int[] m_PrimitiveIndices;
 
         public GCHandle calculatedIndicesHandle;
 
         public MeshTopology topology;
 
-        public PrimitiveCreateContext(int primitiveIndex, int materialCount, string meshName)
-            : base(primitiveIndex, materialCount, meshName)
+        public PrimitiveCreateContext(
+            int meshIndex,
+            int primitiveIndex,
+            int subMeshCount,
+            string meshName
+            )
+            : base(meshIndex, primitiveIndex, subMeshCount, meshName)
         {
-            m_Indices = new int[materialCount][];
+            m_Indices = new int[subMeshCount][];
+            m_PrimitiveIndices = new int[subMeshCount];
         }
 
         public void SetIndices(int subMesh, int[] indices)
@@ -47,9 +54,14 @@ namespace GLTFast
             m_Indices[subMesh] = indices;
         }
 
+        public void SetPrimitiveIndex(int subMesh, int primitiveIndex)
+        {
+            m_PrimitiveIndices[subMesh] = primitiveIndex;
+        }
+
         public override bool IsCompleted => jobHandle.IsCompleted;
 
-        public override async Task<Primitive?> CreatePrimitive()
+        public override async Task<MeshResult?> CreatePrimitive()
         {
             Profiler.BeginSample("CreatePrimitive");
             jobHandle.Complete();
@@ -152,7 +164,7 @@ namespace GLTFast
 
             Profiler.EndSample();
 
-            return new Primitive(msh, m_Materials);
+            return new MeshResult(MeshIndex, m_PrimitiveIndices, m_Materials, msh);
         }
 
         void Dispose()